TAP diagnostic should go into <system-out/>
The default UseCommentsAsTestNames feature is non-standard.
For standard TAP output (http://testanything.org/) like produced by
Bats, comments indicate diagnostic output and without it essential
information about the test case is missing. It should be added to the
report grouped together with the assertion that produced it.

The change needed is fairly trivial. A CLI flag ``--oneAssertionPerTestCase`` is
created to enable the behaviour.